interview_advisor/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.17.1
pillow>=9.0.0
elevenlabs>=0.2.24
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
wave>=0.0.2
flask>=2.0.1
flask-cors>=3.0.10
requests>=2.25.1
numpy>=1.20.0
opencv-python>=4.5.0
mediapipe>=0.8.9
streamlit==1.28.0
pytesseract==0.3.10
langchain==0.0.335
pydub==0.25.1 